+++ tests/Rector/Argument/ArgumentAdderRector/ArgumentAdderRectorTest.php
@@ -1,14 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Dynamic\ArgumentDefaultValueReplacerRector;
+namespace Rector\Tests\Rector\Argument\ArgumentAdderRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-/**
- * @covers \Rector\Rector\Dynamic\ArgumentDefaultValueReplacerRector
- */
-final class ArgumentDefaultValueReplacerRectorTest extends AbstractRectorTestCase
+final class ArgumentAdderRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -22,9 +19,7 @@ final class ArgumentDefaultValueReplacerRectorTest extends AbstractRectorTestCas
     {
         yield [__DIR__ . '/Wrong/wrong.php.inc', __DIR__ . '/Correct/correct.php.inc'];
         yield [__DIR__ . '/Wrong/wrong2.php.inc', __DIR__ . '/Correct/correct2.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong3.php.inc', __DIR__ . '/Correct/correct3.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong4.php.inc', __DIR__ . '/Correct/correct4.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong5.php.inc', __DIR__ . '/Correct/correct5.php.inc'];
+        // yield [__DIR__ . '/Wrong/wrong3.php.inc', __DIR__ . '/Correct/correct3.php.inc'];
     }
 
     protected function provideConfig(): string
