+++ tests/Rector/Argument/ArgumentAdderRector/Correct/correct.php.inc
