+++ tests/Rector/Argument/ArgumentAdderRector/Correct/correct2.php.inc
