+++ tests/Rector/Argument/ArgumentAdderRector/Correct/correct3.php.inc
