+++ tests/Rector/Argument/ArgumentAdderRector/Wrong/wrong2.php.inc
