+++ tests/Rector/Argument/ArgumentAdderRector/Wrong/wrong3.php.inc
