+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Correct/correct.php.inc
