+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Correct/correct2.php.inc
