+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Correct/correct3.php.inc
