+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Correct/correct4.php.inc
