+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Correct/correct5.php.inc
