+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Wrong/wrong.php.inc
