+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Wrong/wrong2.php.inc
