+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Wrong/wrong3.php.inc
