+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Wrong/wrong4.php.inc
