+++ tests/Rector/Argument/ArgumentDefaultValueReplacerRector/Wrong/wrong5.php.inc
