+++ tests/Rector/Argument/ArgumentRemoverRector/Correct/correct.php.inc
