+++ tests/Rector/Argument/ArgumentRemoverRector/Correct/correct2.php.inc
