+++ tests/Rector/Argument/ArgumentRemoverRector/Wrong/wrong.php.inc
