+++ tests/Rector/Argument/ArgumentRemoverRector/Wrong/wrong2.php.inc
